requirements.txt
click
opencv-python
face_recognition
mysql-connector-python
rich
